tests: add xetpriority.test
* resource.c (SYS_FUNC(getpriority), SYS_FUNC(setpriority)): Print
the second syscall argument using %d format.
* tests/xetpriority.c: New file.
* tests/xetpriority.test: New test.
* tests/.gitignore: Add xetpriority.
* tests/Makefile.am (check_PROGRAMS): Likewise.
(TESTS): Add xetpriority.test.

+++ resource.c
@@ -184,7 +184,7 @@ SYS_FUNC(osf_getrusage)
 SYS_FUNC(getpriority)
 {
 	printxval(priorities, tcp->u_arg[0], "PRIO_???");
-	tprintf(", %lu", tcp->u_arg[1]);
+	tprintf(", %d", (int) tcp->u_arg[1]);
 
 	return RVAL_DECODED;
 }
@@ -192,7 +192,7 @@ SYS_FUNC(getpriority)
 SYS_FUNC(setpriority)
 {
 	printxval(priorities, tcp->u_arg[0], "PRIO_???");
-	tprintf(", %lu, %d", tcp->u_arg[1], (int) tcp->u_arg[2]);
+	tprintf(", %d, %d", (int) tcp->u_arg[1], (int) tcp->u_arg[2]);
 
 	return RVAL_DECODED;
 }
